+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Added blocking statuses. If they are missing or failing, they will prevent deploy even if they were reported on any of the commits in the deploy range.
 * Fix shipit.yml updates not being taken into account for the `fetch` command.
 * Fix failing membership webhooks for non fully downcase organization names.
 * Fix status group links not being disabled when the status url is missing (https://github.com/Shopify/shipit-engine/issues/742).